+++ dflib-echarts/src/main/java/org/dflib/echarts/ChartType.java
@@ -4,7 +4,12 @@ package org.dflib.echarts;
  * Enumerates supported chart types.
  */
 public enum ChartType {
-    line, bar, scatter, pie, candlestick, boxplot;
+    line, bar, scatter, pie, candlestick, boxplot,
+
+    /**
+     * @since 2.0.0
+     */
+    heatmap;
 
     /**
      * @deprecated in favor of {@link CoordinateSystemType#isCartesian()}, as many chart types can be plotted over
@@ -23,6 +28,20 @@ public enum ChartType {
                 return false;
             default:
                 throw new UnsupportedOperationException("Unexpected ChartType: " + this);
+        }
+    }
+
+    /**
+     * @since 2.0.0
+     */
+    public boolean supportsDataset() {
+        switch (this) {
+            // TODO: dataset / encode kinda works with heatmaps, but is very unreliable, so should encode data
+            //  per Series for those
+            case heatmap:
+                return false;
+            default:
+                return true;
         }
     }
 }